Move utility mongo method into main mongo class
It's only used by one method, and making it private means I can simplify
without changing an API.
