+++ src/us/kbase/narrativemethodstore/db/mongo/MongoUtils.java
@@ -6,23 +6,19 @@ import java.nio.charset.Charset;
 import java.security.MessageDigest;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.regex.Pattern;
 
 import org.bson.BSONObject;
 import org.bson.LazyBSONList;
 import org.bson.types.BasicBSONList;
-import org.jongo.MongoCollection;
 
 import us.kbase.common.service.UObject;
 import us.kbase.narrativemethodstore.exceptions.NarrativeMethodStoreException;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Lists;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBCollection;
 import com.mongodb.DBCursor;
@@ -53,40 +49,6 @@ public class MongoUtils {
             ret.add((T)value);
         }
         return ret;
-    }
-
-    @SuppressWarnings({ "rawtypes", "unchecked" })
-    public static <KT, VT> Map<KT, VT> getProjection(MongoCollection infos, String whereCondition, 
-            String keySelectField, Class<KT> keyType, String valueSelectField, Class<VT> valueType, 
-            Object... params) throws NarrativeMethodStoreException {
-        List<Map> data = Lists.newArrayList(infos.find(whereCondition, params).projection(
-                "{'" + keySelectField + "':1,'" + valueSelectField + "':1}").as(Map.class));
-        Map<KT, VT> ret = new LinkedHashMap<KT, VT>();
-        for (Map<?,?> item : data) {
-            Object key = getMongoProp(item, keySelectField);
-            if (key == null || !(keyType.isInstance(key)))
-                throw new NarrativeMethodStoreException("Key is wrong: " + key);
-            Object value = getMongoProp(item, valueSelectField);
-            if (value == null)
-                throw new NullPointerException("Value is not defined for selected " +
-                        "field: " + valueSelectField);
-            if (!valueType.isInstance(value))
-                value = UObject.transformObjectToObject(value, valueType);
-            ret.put((KT)key, (VT)value);
-        }
-        return ret;
-    }
-    
-    private static Object getMongoProp(Map<?,?> data, String propWithDots) {
-        String[] parts = propWithDots.split(Pattern.quote("."));
-        Object value = null;
-        for (String part : parts) {
-            if (value != null) {
-                data = (Map<?,?>)value;
-            }
-            value = data.get(part);
-        }
-        return value;
     }
 
     /** Map an object to a MongoDB {@link DBObject}. The object must be serializable by
